歌词相关的 Int -> Long 和 Float -> Double；修改完 offset 及时更新

+++ app/src/main/java/remix/myplayer/helper/MusicServiceRemote.kt
@@ -135,7 +135,7 @@ object MusicServiceRemote {
   }
 
   @JvmStatic
-  fun setProgress(progress: Int) {
+  fun setProgress(progress: Long) {
     service?.setProgress(progress)
   }
